all: Convert score and depth to int32.
No functional changes.
STC
24300 @ 40/15+0.05
8761 - 8711 - 6828
ELO 0.71±3.70

+++ zurichess/uci.go
@@ -210,7 +210,7 @@ func (uci *UCI) go_(line string) error {
 		case "depth":
 			i++
 			d, _ := strconv.Atoi(args[i])
-			uci.timeControl.Depth = d
+			uci.timeControl.Depth = int32(d)
 		}
 	}
